make validate example better
